refactor(types): move schema preview types into its own file and exports types
