+++ packages/@sanity/types/src/schema/index.ts
@@ -1,2 +1,3 @@
 export * from './types'
 export * from './asserters'
+export * from './preview'
